requirements.txt
flask
flask-socketio
python-dotenv
python-telegram-bot==20.5
geopy
requests
eventlet
firebase-admin